backend/requirements.txt
Django==4.2
djangorestframework==3.16.1
django-cors-headers==4.9.0
mysqlclient==2.2.7
python-dotenv==1.1.1
djangorestframework-simplejwt==5.2.2
